+++ tokenizer/gpt2_tokenizer.cc
@@ -17,7 +17,6 @@
 #include <codecvt>
 #include <mutex>
 
-#include "gpt2_tokenizer.hpp"
 #include "nlohmann/json.hpp"
 #include "kernels/kernels.h"
 #include "kernels/string_common.h"
@@ -459,6 +458,27 @@ bool IsUnicodeSpace(char32_t ch) {
 bool IsEmptyUString(const ustring& str) {
   return std::all_of(str.begin(), str.end(), [](char32_t ch) { return IsUnicodeSpace(ch); });
 }
+
+struct KernelBpeTokenizer : BaseKernel {
+  KernelBpeTokenizer(OrtApi api, const OrtKernelInfo* info);
+  void Compute(OrtKernelContext* context);
+
+ private:
+  std::vector<int64_t> Tokenize(const ustring& input, int64_t max_length);
+
+  int64_t padding_length_;
+  std::list<int> byte_list_;
+  std::shared_ptr<VocabData> bbpe_tokenizer_;
+};
+
+struct CustomOpBpeTokenizer : Ort::CustomOpBase<CustomOpBpeTokenizer, KernelBpeTokenizer> {
+  void* CreateKernel(OrtApi api, const OrtKernelInfo* info) const;
+  const char* GetName() const;
+  size_t GetInputTypeCount() const;
+  ONNXTensorElementDataType GetInputType(size_t index) const;
+  size_t GetOutputTypeCount() const;
+  ONNXTensorElementDataType GetOutputType(size_t index) const;
+};
 
 KernelBpeTokenizer::KernelBpeTokenizer(OrtApi api, const OrtKernelInfo* info)
     : BaseKernel(api, info) {
